+++ frontend/app/src/main/java/com/example/gestionrecetaschef/domain/modelo/OpinionRequest.kt
@@ -6,7 +6,7 @@ data class OpinionRequest(
 
     val comentario: String,
 
-    val puntuacion: Int,
+    val puntuacion: Double,
 
     val veces_preparada: Int
 )
